+++ poker-hands/project.tests/CardTests.cs
@@ -1,4 +1,4 @@
-using katacombs;
+using Project;
 using NUnit.Framework;
 
 namespace Tests
