+++ poker-hands/project/Suit.cs
@@ -1,4 +1,4 @@
-namespace katacombs
+namespace Project
 {
     public enum Suit
     {
